Enhance Blockaid transaction input resolvers with new utility function
* Introduce getCompiledTxsForBlockaidInput utility to streamline transaction compilation across EVM, Solana, and Sui resolvers.
* Refactor getSolanaBlockaidTxScanInput and getSuiBlockaidTxScanInput to utilize the new utility for improved clarity and maintainability.
* Add metadata field to transaction input structures for better data handling.
