+++ core/chain/security/blockaid/tx/scan/input/resolvers/solana.ts
@@ -1,11 +1,8 @@
-import { getCoinType } from '@core/chain/coin/coinType'
-import { getKeysignSwapPayload } from '@core/mpc/keysign/swap/getKeysignSwapPayload'
-import { getKeysignTwPublicKey } from '@core/mpc/keysign/tw/getKeysignTwPublicKey'
-import { getSolanaSendTxInputData } from '@core/mpc/keysign/txInputData/solana/send'
 import { assertField } from '@lib/utils/record/assertField'
 
 import { OtherChain } from '../../../../../../Chain'
 import { decodeTx } from '../../../../../../tx/decode'
+import { getCompiledTxsForBlockaidInput } from '../../utils/getCompiledTxsForBlockaidInput'
 import { BlockaidTxScanInputResolver } from '../resolver'
 
 export const getSolanaBlockaidTxScanInput: BlockaidTxScanInputResolver<
@@ -13,52 +10,24 @@ export const getSolanaBlockaidTxScanInput: BlockaidTxScanInputResolver<
 > = ({ payload, walletCore, chain }) => {
   const coin = assertField(payload, 'coin')
 
-  const swapPayload = getKeysignSwapPayload(payload)
-
-  if (swapPayload) {
-    return null
-  }
-
-  const txInputData = getSolanaSendTxInputData({
-    keysignPayload: payload,
+  const transactions = getCompiledTxsForBlockaidInput({
+    payload,
     walletCore,
-  })
-
-  const publicKeyData = getKeysignTwPublicKey(payload)
-  const publicKey = walletCore.PublicKey.createWithData(
-    publicKeyData,
-    walletCore.PublicKeyType.ed25519
+  }).map(
+    tx =>
+      decodeTx({
+        chain,
+        compiledTx: tx,
+      }).encoded
   )
-
-  const allSignatures = walletCore.DataVector.create()
-  const publicKeys = walletCore.DataVector.create()
-
-  allSignatures.add(Buffer.from('0'.repeat(128), 'hex'))
-  publicKeys.add(publicKey.data())
-
-  const coinType = getCoinType({
-    chain,
-    walletCore,
-  })
-
-  const compiledTx = walletCore.TransactionCompiler.compileWithSignatures(
-    coinType,
-    txInputData,
-    allSignatures,
-    publicKeys
-  )
-
-  const { encoded } = decodeTx({
-    chain,
-    compiledTx,
-  })
 
   return {
     chain: 'mainnet',
     options: ['validation'],
     account_address: coin.address,
     encoding: 'base58',
-    transactions: [encoded],
+    transactions,
     method: 'signAndSendTransaction',
+    metadata: {},
   }
 }
