+++ core/chain/security/blockaid/tx/scan/input/resolvers/sui.ts
@@ -1,10 +1,8 @@
-import { getCoinType } from '@core/chain/coin/coinType'
-import { getKeysignTwPublicKey } from '@core/mpc/keysign/tw/getKeysignTwPublicKey'
-import { getSuiTxInputData } from '@core/mpc/keysign/txInputData/sui'
 import { assertField } from '@lib/utils/record/assertField'
 
 import { OtherChain } from '../../../../../../Chain'
 import { decodeTx } from '../../../../../../tx/decode'
+import { getCompiledTxsForBlockaidInput } from '../../utils/getCompiledTxsForBlockaidInput'
 import { BlockaidTxScanInputResolver } from '../resolver'
 
 export const getSuiBlockaidTxScanInput: BlockaidTxScanInputResolver<
@@ -12,45 +10,24 @@ export const getSuiBlockaidTxScanInput: BlockaidTxScanInputResolver<
 > = ({ payload, walletCore }) => {
   const coin = assertField(payload, 'coin')
 
-  const [txInputData] = getSuiTxInputData({
-    keysignPayload: payload,
-    walletCore,
-    chain: OtherChain.Sui,
-  })
-
-  const publicKeyData = getKeysignTwPublicKey(payload)
-  const publicKey = walletCore.PublicKey.createWithData(
-    publicKeyData,
-    walletCore.PublicKeyType.ed25519
-  )
-
-  const allSignatures = walletCore.DataVector.create()
-  const publicKeys = walletCore.DataVector.create()
-
-  allSignatures.add(Buffer.from('0'.repeat(128), 'hex'))
-  publicKeys.add(publicKey.data())
-
-  const coinType = getCoinType({
-    chain: OtherChain.Sui,
+  const compiledTxs = getCompiledTxsForBlockaidInput({
+    payload,
     walletCore,
   })
 
-  const compiledTx = walletCore.TransactionCompiler.compileWithSignatures(
-    coinType,
-    txInputData,
-    allSignatures,
-    publicKeys
+  const [transaction] = compiledTxs.map(
+    compiledTx =>
+      decodeTx({
+        chain: OtherChain.Sui,
+        compiledTx,
+      }).unsignedTx
   )
-
-  const { unsignedTx } = decodeTx({
-    chain: OtherChain.Sui,
-    compiledTx,
-  })
 
   return {
     chain: 'mainnet',
     options: ['validation'],
     account_address: coin.address,
-    transaction: unsignedTx,
+    transaction,
+    metadata: {},
   }
 }
